Fixed #60
I added a link at the bottom of the page, along with a bit of text
explaining what the link is. I will be using this link to make a new
issue, to add a form on the page to make links automatically. I feel
that this is preferable to a link, particularly if I can get it without
a github account.